+++ src/@cdk/components/compartilhamento/cdk-compartilhamento-form/cdk-compartilhamento-form.module.ts
@@ -14,6 +14,17 @@ import {UsuarioService} from '@cdk/services/usuario.service';
 import {CdkCompartilhamentoFormComponent} from './cdk-compartilhamento-form.component';
 import {CdkUsuarioAutocompleteModule} from '@cdk/components/usuario/cdk-usuario-autocomplete/cdk-usuario-autocomplete.module';
 import {CdkUsuarioGridsearchModule} from '@cdk/components/usuario/cdk-usuario-autocomplete/cdk-usuario-gridsearch/cdk-usuario-gridsearch.module';
+import {SetorService} from '@cdk/services/setor.service';
+import {CdkSetorAutocompleteModule} from '@cdk/components/setor/cdk-setor-autocomplete/cdk-setor-autocomplete.module';
+import {CdkSetorGridsearchModule} from '@cdk/components/setor/cdk-setor-autocomplete/cdk-setor-gridsearch/cdk-setor-gridsearch.module';
+import {
+    CdkModalidadeCompartilhamentoAutocompleteModule
+} from "../../modalidade-compartilhamento/cdk-modalidade-compartilhamento-autocomplete/cdk-modalidade-compartilhamento-autocomplete.module";
+import {
+    CdkModalidadeCompartilhamentoGridsearchModule
+} from "../../modalidade-compartilhamento/cdk-modalidade-compartilhamento-autocomplete/cdk-modalidade-compartilhamento-gridsearch/cdk-modalidade-compartilhamento-gridsearch.module";
+import {MatSelectModule} from "@angular/material/select";
+import {ModalidadeCompartilhamentoService} from "../../../services/modalidade-compartilhamento.service";
 
 @NgModule({
     declarations: [
@@ -32,10 +43,18 @@ import {CdkUsuarioGridsearchModule} from '@cdk/components/usuario/cdk-usuario-au
         CdkUsuarioAutocompleteModule,
         CdkUsuarioGridsearchModule,
 
+        CdkSetorAutocompleteModule,
+        CdkSetorGridsearchModule,
+
         CdkSharedModule,
+        CdkModalidadeCompartilhamentoAutocompleteModule,
+        CdkModalidadeCompartilhamentoGridsearchModule,
+        MatSelectModule,
     ],
     providers: [
         UsuarioService,
+        SetorService,
+        ModalidadeCompartilhamentoService
     ],
     exports: [
         CdkCompartilhamentoFormComponent