+++ src/@cdk/data-sources/modalidade-compartilhamento-data-source.ts
@@ -1,0 +1,17 @@
+import {CollectionViewer, DataSource} from '@angular/cdk/collections';
+import {Observable} from 'rxjs';
+import {ModalidadeCompartilhamento} from '@cdk/models';
+
+export class ModalidadeCompartilhamentoDataSource extends DataSource<ModalidadeCompartilhamento> {
+
+    public constructor(private compartilhamentoes$: Observable<ModalidadeCompartilhamento[]>) {
+        super();
+    }
+
+    public connect(collectionViewer: CollectionViewer): Observable<ModalidadeCompartilhamento[]> {
+        return this.compartilhamentoes$;
+    }
+
+    public disconnect(collectionViewer: CollectionViewer): void {
+    }
+}