+++ src/app/main/apps/documento-avulso/widget/colaborador/widget-documento-avulso.component.ts
@@ -55,7 +55,7 @@ export class WidgetDocumentoAvulsoComponent implements OnInit {
      */
     ngOnInit(): void {
         this._documentoAvulsoService.count(
-            `{"usuarioResponsavel.id": "eq:${this._profile.id}", "dataHoraResposta": "isNull","dataHoraRemessa": "isNotNull"}`)
+            `{"usuarioResponsavel.id": "eq:${this._profile.id}", "dataHoraResposta": "isNull","dataHoraRemessa": "isNotNull", "dataHoraEncerramento": "isNull"}`)
             .pipe(
                 catchError(() => of([]))
             ).subscribe(