+++ CHANGELOG.md
@@ -856,3 +856,4 @@
 * Correção para buscar o provider de processo com 15 dígitos
 * Melhoria inclusão do Copiar Bookmark para clipboard
 * Melhoria inclusão do Copiar número do documento juntado para o clipboard
+* Melhoria inclusão do Compartilhar Tarefa para Setor